+++ packages/docs/src/stories/Tooltip.stories.tsx
@@ -1,0 +1,29 @@
+import type { StoryObj, Meta } from '@storybook/react'
+import { Box, Text, Tooltip, TooltipProps } from '@lpatestorg/react'
+
+export default {
+  title: 'Form/Tooltip',
+  component: Tooltip,
+  args: {
+    children: <Text>Testando elemento Tooltip</Text>,
+    message: 'Tooltip message',
+  },
+  argTypes: {
+    children: {
+      control: {
+        type: null,
+      },
+    },
+  },
+  decorators: [
+    (Story) => {
+      return (
+        <Box css={{ display: 'flex', flexDirection: 'row', gap: '$2' }}>
+          {Story()}
+        </Box>
+      )
+    },
+  ],
+} as Meta<TooltipProps>
+
+export const Primary: StoryObj<TooltipProps> = {}