+++ packages/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @lpatestorg/react
 
+## 4.0.0
+
+### Major Changes
+
+- New components: Toast and Tooltip
+
+## 3.0.0
+
+### Major Changes
+
+- New Components Toast and ToolTip
+
 ## 2.1.0
 
 ### Minor Changes